First attempt
