--- js/file_management/errors.js
@@ -1,13 +1,0 @@
-const ErrorCode = {
-    NoAccess: 0,
-    FileExists: 1,
-    FileNotFound: 2,
-    DirectoryExists: 3,
-    DirectoryNotFound: 4,
-    WrongFileExtension: 5,
-    MaxFileSizeExceeded: 6,
-    InvalidSymbols: 7,
-    Other: 32767
-};
-
-export default ErrorCode;
