+++ js/ui/file_manager/ui.file_manager.messages.js
@@ -1,5 +1,5 @@
 import messageLocalization from '../../localization/message';
-import ErrorCode from '../../file_management/errors';
+import { ErrorCode } from '../../file_management/file_system_error';
 
 export const FileManagerMessages = {
     get: (errorId, args) => {
@@ -25,5 +25,3 @@ export const FileManagerMessages = {
         return messageLocalization.format('dxFileManager-errorDefault');
     }
 };
-
-export { ErrorCode };
